+++ android/src/main/java/dev/robingenz/capacitor/appupdate/AppUpdateInfoResultCallback.java
@@ -1,0 +1,9 @@
+package dev.robingenz.capacitor.appupdate;
+
+import com.getcapacitor.JSObject;
+import com.google.android.play.core.appupdate.AppUpdateInfo;
+
+public interface AppUpdateInfoResultCallback {
+    void onDone(AppUpdateInfo info);
+    void onError(String message);
+}
